Fix Breakpoints in JS implementation
Unfortunately, it appears to be the case that there's inevitably a
noticeable lag when a visible change happens onMount. This greatly
complicated things because the site's responsiveness (in the design
sense) should be visible immediately when the site is loaded, not after
a delay. I tried things as desperate as putting properties on `console`
in app.html to "smuggle" the viewport width into Svelte before onMount
could happen, but even then, the lag was too much. The best I could
seem to do was to make a client-side hook, and make a component that
uses CSS until the breakpoints are ready.